mvp for replacing service --status-all with systemctl is-active to ensure app is stopped
fixes https://github.com/Sonarr/Sonarr/issues/7960#issue-3209801021

service --status-all does not return a status for sonarr unit. 

also see 7960 for discussion on if stopping is required at all.  Upgrade worked for me despite this bug.  If so, we can simply remove this check and restart the unit at the end.